+++ scripts/train.py
@@ -6,7 +6,7 @@ import wandb
 from datasets import concatenate_datasets
 
 from arcsf import ExperimentConfig, load_model_and_tokenizer, load_trainer
-from arcsf.config.experiment_config import EXPERIMENT_CONFIG_DIR
+from arcsf.config.experiment import EXPERIMENT_CONFIG_DIR
 from arcsf.data.tofu import load_tofu
 from arcsf.utils import seed_everything
